classes/cargo: create linker wrapper script
Create a linker wrapper script to handle the fact that Rust really
expects the linker to be a single binary and Yocto really wants to
supply it with arguments. fixes #76. Allows us to supply arguments
easily to the linker instead of gaming with RUSTFLAGS or the cargo
config via rustflags. Effectively redoes
4e1bda643ef7896bdaa77261a3abd8ce312e41c9 and
5afc0cc471eef5c18bbce9bc54aed6110e7c55eb in what I hope works better
going forward.